refactor(local-ci): extract github_workflows.py (R2-1 ninth cut)
Move the GitHub Actions workflow dispatch cluster out of local_ci.py
into a sibling github_workflows.py module:

- 3 constants: GITHUB_ACTIONS_DEFAULTS, BUILTIN_GITHUB_WORKFLOWS,
  REPO_VARIABLE_FALLBACKS
- 11 resolver functions: github_actions_settings_for_display,
  resolve_github_actions_settings, normalize_runs_on_json,
  resolve_workflow_runner_selector_json,
  resolve_workflow_dispatch_field_values,
  repo_variable_name_for_workflow_field,
  resolve_default_provider_for_workflow,
  resolve_workflow_field_value_and_source,
  resolve_workflow_dispatch_defaults, summarize_workflow_provider_defaults,
  resolve_cli_dispatch_field_values

Pure functions — they accept a config dict + CLI args, return derived
dicts/strings. No I/O, no subprocess, no GitHub API calls. The cloud-
dispatch orchestrator in local_ci.py still owns the actual subprocess
gh-api invocation.

- local_ci.py: 10,803 → 10,514 lines (-289 net).
- github_workflows.py: 333 lines.
- Re-exported from local_ci.py for backward compat.

Tests: 213/213 local_ci tests pass.

Running R2-1 total: 9 modules extracted (state_paths, normalize,
git_helpers, io_utils, footprint, provenance, job_queue, targets,
github_workflows). local_ci.py shrunk 11,285 → 10,514 lines
(-771, -6.8%).
